first test.

+++ pkg/entities/DataCollected.go
@@ -5,7 +5,7 @@ type DataCollected struct {
 	Speedup              float64
 	Efficiency           float64
 	ExecutionTime        int64
-	OverHead             float64
+	OverHead             int64
 	IdleThreadTimeMedian int64
 	IsSingleThread       bool
 	FullExecutionTime    int64
